Updated
I added the fireworks and the lyrics for the second chorus, to do this I took the same lyrics and fireworks from the first chorus and pasted them at the bottom of the code. I think changed the timing so that they would appear and dissapear at the right time.